requirements.txt
streamlit
requests
pandas
numpy
gensim
spacy
newsapi-python
scikit-learn